FAD-5629 Renamed action helper from localFileParseRequest to csvFileParseRequest
